+++ C++/dij.cpp
