+++ Java/GalToLit.java
@@ -1,0 +1,14 @@
+/* Try This 1-1
+This Program converts gallons to liters.
+Call this program GalTo Lit.java.
+*/
+class GalToLit{
+public static void main(String args[])
+{
+double gallons;  //hold the number of gallons
+double liters;   //holds conversion to liters
+gallons =10; //start  with 10 gallons
+liters = gallons + 3.7854;  //converts to gallons
+System.out.println(gallons +" "+ "gallons is " +liters + "liters .");
+}
+}
