+++ Java/Stringoperaations.java
