+++ Python/average_list.py
