--- helloworld.java
@@ -1,6 +1,0 @@
-public class helloworld{
-public static void main(String args[])
-{
-System.out.println("Hello World");
-}
-}
